schedprice/requirements.txt
pyyaml
pika
psycopg2
